Add user data streaming functionality and update README.md

+++ python-generators-0x00/seed.py
@@ -64,7 +64,7 @@ def create_table(connection):
                 user_id CHAR(36) PRIMARY KEY,
                 name VARCHAR(255) NOT NULL,
                 email VARCHAR(255) NOT NULL UNIQUE,
-                age DECIMAL(5, 2) NOT NULL
+                age TINYINT NOT NULL
             )
         """)
         print("Table created")
